requirements.txt
openai
python-dotenv
notion_client
requests
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib